requirements.txt
# Core Django & DRF
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1

# Multi-tenancy
django-tenants==3.6.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Caching & Task Queue
redis==5.0.1
django-redis==5.4.0
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Azure Services
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-communication-email==1.0.0
azure-communication-sms==1.0.1
opencensus-ext-azure==1.1.13

# Payments
requests==2.31.0
cryptography==42.0.0

# Media & Files
Pillow==10.2.0
python-magic==0.4.27

# Utils
python-dotenv==1.0.1
python-dateutil==2.8.2
pytz==2024.1

# Testing
pytest==8.0.0
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.6.0

# Code Quality
black==24.1.1
flake8==7.0.0
isort==5.13.2

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring
sentry-sdk==1.40.0
